refactor(planning_data_analyzer): organize EPDMS metric files
