+++ planning/autoware_planning_data_analyzer/src/metrics/deviation_metrics.cpp
@@ -14,7 +14,7 @@
 
 #include "deviation_metrics.hpp"
 
-#include "metric_utils.hpp"
+#include "geometry/metric_utils.hpp"
 
 #include <autoware_utils_geometry/geometry.hpp>
