+++ planning/autoware_planning_data_analyzer/src/metrics/epdms/subscores/history_comfort.cpp
@@ -14,7 +14,7 @@
 
 #include "history_comfort.hpp"
 
-#include "metric_utils.hpp"
+#include "../../geometry/metric_utils.hpp"
 
 #include <autoware_utils_math/normalization.hpp>
