+++ planning/autoware_planning_data_analyzer/src/metrics/epdms/subscores/no_at_fault_collision.hpp
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef METRICS__TTC_WITHIN_BOUND_HPP_
-#define METRICS__TTC_WITHIN_BOUND_HPP_
+#ifndef METRICS__EPDMS__SUBSCORES__NO_AT_FAULT_COLLISION_HPP_
+#define METRICS__EPDMS__SUBSCORES__NO_AT_FAULT_COLLISION_HPP_
 
-#include "../data_types.hpp"
+#include "../../../data_types.hpp"
 
 #include <autoware/route_handler/route_handler.hpp>
 #include <autoware_vehicle_info_utils/vehicle_info.hpp>
@@ -29,7 +29,7 @@ namespace autoware::planning_data_analyzer::metrics
 
 using autoware::route_handler::RouteHandler;
 
-struct TTCWithinBoundResult
+struct NoAtFaultCollisionResult
 {
   double score{0.0};
   bool available{false};
@@ -37,7 +37,7 @@ struct TTCWithinBoundResult
   double infraction_time_s{std::numeric_limits<double>::infinity()};
 };
 
-TTCWithinBoundResult calculate_ttc_within_bound(
+NoAtFaultCollisionResult calculate_no_at_fault_collision(
   const autoware_planning_msgs::msg::Trajectory & trajectory,
   const std::shared_ptr<PredictedObjects> & objects,
   const autoware::vehicle_info_utils::VehicleInfo & vehicle_info,
@@ -45,4 +45,4 @@ TTCWithinBoundResult calculate_ttc_within_bound(
 
 }  // namespace autoware::planning_data_analyzer::metrics
 
-#endif  // METRICS__TTC_WITHIN_BOUND_HPP_
+#endif  // METRICS__EPDMS__SUBSCORES__NO_AT_FAULT_COLLISION_HPP_
