+++ planning/autoware_planning_data_analyzer/src/metrics/geometry/metric_utils.cpp
