+++ planning/autoware_planning_data_analyzer/test/metrics/test_extended_comfort.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../../src/metrics/extended_comfort.hpp"
+#include "../../src/metrics/epdms/subscores/extended_comfort.hpp"
 
 #include <gtest/gtest.h>
