+++ planning/autoware_planning_data_analyzer/test/metrics/test_traffic_light_compliance.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../../src/metrics/traffic_light_compliance.hpp"
+#include "../../src/metrics/epdms/subscores/traffic_light_compliance.hpp"
 
 #include <gtest/gtest.h>
